seperated gamescreen and menuscreen into to classes
